front-end | Moving files in folder + reservation page improvements
